requirements.txt
python-socketio==4.5.1